Revert "fix: handle selection mode in search context"
