+++ src/components/SelectionListWithModal/index.tsx
@@ -1,21 +1,21 @@
 import {useIsFocused} from '@react-navigation/native';
 import type {ForwardedRef} from 'react';
-import React, {forwardRef, useMemo, useState} from 'react';
+import React, {forwardRef, useEffect, useRef, useState} from 'react';
 import {CheckSquare} from '@components/Icon/Expensicons';
 import MenuItem from '@components/MenuItem';
 import Modal from '@components/Modal';
 import SelectionList from '@components/SelectionListWithSections';
 import type {ListItem, SelectionListHandle, SelectionListProps} from '@components/SelectionListWithSections/types';
-import useHandleSelectionMode from '@hooks/useHandleSelectionMode';
 import useLocalize from '@hooks/useLocalize';
 import useMobileSelectionMode from '@hooks/useMobileSelectionMode';
 import useResponsiveLayout from '@hooks/useResponsiveLayout';
-import {turnOnMobileSelectionMode} from '@libs/actions/MobileSelectionMode';
+import {turnOffMobileSelectionMode, turnOnMobileSelectionMode} from '@libs/actions/MobileSelectionMode';
 import CONST from '@src/CONST';
 
 type SelectionListWithModalProps<TItem extends ListItem> = SelectionListProps<TItem> & {
     turnOnSelectionModeOnLongPress?: boolean;
     onTurnOnSelectionMode?: (item: TItem | null) => void;
+    isSelected?: (item: TItem) => boolean;
     isScreenFocused?: boolean;
 };
 
@@ -42,20 +42,52 @@ function SelectionListWithModal<TItem extends ListItem>(
     const isFocused = useIsFocused();
 
     const isMobileSelectionModeEnabled = useMobileSelectionMode();
+    // Check if selection should be on when the modal is opened
+    const wasSelectionOnRef = useRef(false);
+    // Keep track of the number of selected items to determine if we should turn off selection mode
+    const selectionRef = useRef(0);
 
-    const selectedItems = useMemo(
-        () =>
+    useEffect(() => {
+        // We can access 0 index safely as we are not displaying multiple sections in table view
+        const selectedItems =
             selectedItemsProp ??
             sections[0].data.filter((item) => {
                 if (isSelected) {
                     return isSelected(item);
                 }
                 return !!item.isSelected;
-            }),
-        [isSelected, sections, selectedItemsProp],
-    );
+            });
+        selectionRef.current = selectedItems.length;
 
-    useHandleSelectionMode(selectedItems);
+        if (!isSmallScreenWidth) {
+            if (selectedItems.length === 0 && isMobileSelectionModeEnabled) {
+                turnOffMobileSelectionMode();
+            }
+            return;
+        }
+        if (!isFocused) {
+            return;
+        }
+        if (!wasSelectionOnRef.current && selectedItems.length > 0) {
+            wasSelectionOnRef.current = true;
+        }
+        if (selectedItems.length > 0 && !isMobileSelectionModeEnabled) {
+            turnOnMobileSelectionMode();
+        } else if (selectedItems.length === 0 && isMobileSelectionModeEnabled && !wasSelectionOnRef.current) {
+            turnOffMobileSelectionMode();
+        }
+        // eslint-disable-next-line react-compiler/react-compiler, react-hooks/exhaustive-deps
+    }, [sections, selectedItemsProp, isMobileSelectionModeEnabled, isSmallScreenWidth, isSelected, isFocused]);
+
+    useEffect(
+        () => () => {
+            if (selectionRef.current !== 0) {
+                return;
+            }
+            turnOffMobileSelectionMode();
+        },
+        [],
+    );
 
     const handleLongPressRow = (item: TItem) => {
         // eslint-disable-next-line @typescript-eslint/prefer-nullish-coalescing
